Move profile folder into ui folder
